Fix paths that broke in windows launcher sources when directories were moved around.
