+++ .gitignore
@@ -27,13 +27,13 @@
 /nbi/engine/native/*/*/dist/
 /nb-javac/
 /java.source.nbjavac/test/test-nb-javac/nbproject/private/
-/apisupport.harness/windows-launcher-src/*.exe
-/apisupport.harness/windows-launcher-src/*.res
-/ide/launcher/windows/*.exe
-/ide/launcher/windows/*.res
-/o.n.bootstrap/launcher/windows/*.dll
-/o.n.bootstrap/launcher/windows/*.exe
-/o.n.bootstrap/launcher/windows/*.res
+/harness/apisupport.harness/windows-launcher-src/*.exe
+/harness/apisupport.harness/windows-launcher-src/*.res
+/nb/ide.launcher/windows/*.exe
+/nb/ide.launcher/windows/*.res
+/platform/o.n.bootstrap/launcher/windows/*.dll
+/platform/o.n.bootstrap/launcher/windows/*.exe
+/platform/o.n.bootstrap/launcher/windows/*.res
 
 
 # OS generated files #
